requirements.txt
ansi2html==1.8.0
black==23.7.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.6
coverage==7.3.0
dash==2.12.0
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.7
filelock==3.12.2
Flask==2.2.5
gunicorn==21.2.0
identify==2.5.26
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
plotly==5.16.0
pluggy==1.2.0
pre-commit==3.3.3
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
ruff==0.0.284
six==1.16.0
tenacity==8.2.3
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.3
Werkzeug==2.2.3
